setings: set up for translations
Use "ubuntu-system-settings" for the domain, better to namespace to avoid
name conflicts since all the translation are installed in the same directory.

The po.pro will make the translation be compiled and installed on build,
there is also a "make pot" target to refresh the template.

This commit also include a first translation (fr.po) which was used 
to test that things are working correctly.