license: implement the "software licenses" section subpanel
The current approch is to list all the packages installing a copyright and
build a model from that. When selecting one of them in the list we go to a
new page which displays the content of the copyright.
That might need some tweaking later to include licences from the container