battery: don't animate bleutooth switch when opening the page
This happened because the switch's value is not set on initialization, but when
the bluetooth indicator service reports its state a few ms after opening the
panel. The switch always animates between its states, with no apparent way to
turn it off.

This fix puts the switch control into a Loader, which loads it as soon as the
bluetooth-enabled action is available.