MainView: return success if a plugin has been loaded
The loadPluginByName() function was not returning anything in the case where a
plugin could be loaded but didn't provide a page component (like Online
Accounts is doing). In this case probably JS returns just false.
Make sure that the function returns a value in all cases: false if the plugin
cannot be loaded, true if the plugin was loaded successfully (even if it
doesn't provide a page component).