bluetooth: Add Appearance characteristic to type conversion
The Bluetooth LE devices use the GATT Appearance characteristic
instead of the CoD field to introduce themselves. So far the
Bluetooth plugin for u-s-s have not read it and as a result
the BLE devices were invisible on the found devices list. This
change allows them to be visible as right now the type of
device is also guessed based on the Appearance property.